requirements.txt
# Telegram Bot Framework
python-telegram-bot>=20.0

# HTTP Client for API calls
httpx>=0.24.0

# Async Task Scheduling
apscheduler>=3.10.0

# Environment Variables
python-dotenv>=1.0.0

# Database & ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter
alembic>=1.13.0         # Database migrations

# Caching & Session Management
redis>=5.0.0

# Monitoring & Error Tracking
prometheus-client>=0.19.0
sentry-sdk>=1.40.0

# Enhanced Logging & Metrics
structlog>=23.0.0

# Data Validation
pydantic>=2.5.0

# Async Database Operations
asyncpg>=0.29.0         # Async PostgreSQL driver
databases>=0.8.0        # Async database interface

# Testing & Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0      # Coverage reporting
pytest-xdist>=3.5.0    # Parallel test execution
pytest-mock>=3.12.0    # Enhanced mocking
pytest-timeout>=2.2.0  # Test timeouts
faker>=20.1.0           # Test data generation

# HTTP Server & Monitoring
aiohttp>=3.9.0          # HTTP server for monitoring endpoints
psutil>=5.9.0           # System resource monitoring

# Production Utilities
uvloop>=0.19.0          # High-performance event loop
gunicorn>=21.2.0        # WSGI HTTP Server